regrouped text components
